m260b/requirements.txt
distribute 
virtualenv 
setuptools 
pip 
biopython 
Cython
guppy 
mmh3 
nose
numpy 
pandas 
patsy
pyfasta 
pysam 
python-dateutil 
pytz
PyVCF 
scikit-learn 
scipy
six 
statsmodels 
wsgiref 

